+++ app/src/main/java/de/foodbro/app/model/PreparationStep.kt
@@ -9,8 +9,8 @@ data class PreparationStep @JvmOverloads constructor(
 
     var pos: Int = 0,
 
-    var recipeId: Int = 0,
+    var recipeId: Long = 0,
 
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0
+    var id: Long = 0
 )
